MODIFY: 데코레이터 분리
