+++ users/views.py
@@ -10,6 +10,7 @@ from django.views           import View
 from users.models           import User, Gender
 from users.validations      import Validation
 from my_settings            import SECRET_KEY
+from decorators             import validate_login
 
 class SignupView(View):
     def post(self, request):
@@ -92,7 +93,7 @@ class LoginView(View):
             return JsonResponse({'MESSAGE':'NO_BODY'}, status=400)
 
 class UserView(View):
-    @Validation.validate_login
+    @validate_login
     def get(self, request):
         user = request.account
         user_info = {
